requirements.txt
aiohttp==3.6.1
PasteDeploy==1.5.2
gunicorn==19.9.0
aioredis==1.2.0
lxml
arq==0.18
pyyaml==5.1.2
